+++ CHANGELOG.md
@@ -139,7 +139,9 @@ Especially the following are now provided as composables:
   - The `range` property was removed in favor of `type="range"` (datetime ranges) and `type="range-date"` (date only ranges).
   - The `lang` property was replaced with the `locale` property.
   - The `formatter` property was removed.
-  
+- `NcSelect`
+  - `user-select` property was removed, instead just use the `NcSelectUsers` component
+
 ### 🚀 Enhancements
 * Allow writing components using Typescript and provide type definitions for `NcButton` [\#4525](https://github.com/nextcloud-libraries/nextcloud-vue/pull/4525) \([susnux](https://github.com/susnux)\)